+++ packages/elemix/index.ts
@@ -2,3 +2,4 @@ export * from './src/component/Component';
 export { defineComponent } from './src/runtime/dom';
 export * from './src/utilities';
 export { tpl } from './src/tpl';
+export { createApp } from './src/createApp';